backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
sqlalchemy==2.0.29
pydantic==2.6.4
pytest==8.3.3
werkzeug==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
